% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data1.R
\docType{data}
\name{data1}
\alias{data1}
\title{Simulated data with 50 covariates}
\format{An object of class \code{data.frame} with 100 rows and 51 columns.}
\usage{
data(data1)
}
\description{
data1 artificially generated \code{dataframe} with n=100 and p=50. Missing values using MAR (missing at random) mechanism are artificially generated in 10 covariates.
}
\examples{
data(data1)
}
\keyword{datasets}
