% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2.R
\docType{data}
\name{data2}
\alias{data2}
\title{Simulated data with 200 covariates}
\format{An object of class \code{data.frame} with 100 rows and 201 columns.}
\usage{
data(data2)
}
\description{
data2 artificially generated \code{dataframe} with n=100 and p=200. Missing values using MAR (missing at random) mechanism are artificially generated in 10 covariates.
}
\examples{
data(data2)
}
\keyword{datasets}
